[ref] Clean-coder refactored CreateQuiz and Play
