+++ frontend/src/App.js
@@ -5,10 +5,10 @@ import { QuizProvider } from './context/QuizContext';
 import { ScoresProvider } from './context/ScoresContext';
 import { SearchProvider } from './context/SearchContext';
 import { ThemeProvider } from './context/ThemeContext';
-import CreateQuiz from './pages/CreateQuiz';
+import CreateQuiz from './pages/create';
 import Home from './pages/Home';
 import NotFound from './pages/NotFound';
-import Play from './pages/Play';
+import Play from './pages/play';
 import QuizDetail from './pages/QuizDetail';
 import Results from './pages/Results';
 import Review from './pages/Review';
